pyproject.toml
[project]
name = "photo-book-creator"
version = "0.1.0"
description = "Local photo book creation tools"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "Pillow>=10.0.0",
  "python-multipart>=0.0.9",
  "uvicorn>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "httpx>=0.27.0",
]

[project.scripts]
photobook-thumbnails = "photobook.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.pyright]
extraPaths = ["src"]
